# internal/management/openapi/openapi.yaml
components:
...
  schemas:
...
    Alarm:
      properties:
        actorId:
          type: string
        actorType:
          type: string
        alarmId:
          type: string
        dueTime:
          format: date-time
          type: string
        interval:
          description: The repeat interval, for repeating alarms
          type: string
        leaseExpiration:
          description: When the lease expires, present only when leased
          format: date-time
          type: string
        leaseHostId:
          description: The host holding the lease, which is the live host the alarm's
            actor is placed on; omitted when not leased, or when the actor has no
            placement on a live host
          type: string
        leased:
          description: True when a host currently holds a lease on the alarm
          type: boolean
        name:
          type: string
        ttl:
          description: When a repeating alarm stops repeating, omitted when unset
          format: date-time
          type: string
      required:
      - actorId
      - actorType
      - alarmId
      - dueTime
      - leased
      - name
      type: object
...
    Error:
      description: The body of every error response.
      properties:
        code:
          description: |-
            A machine-readable error code
            See the API description for the meaning of each code
          enum:
          - badRequest
          - unauthorized
          - forbidden
          - notFound
          - notApplicable
          - methodNotAllowed
          - exclusiveLeaseHeld
          - lastServer
          - hostUnavailable
          - hostReattached
          - noHostsReachable
          - eventHistoryDisabled
          - payloadTooLarge
          - timeout
          - internal
          type: string
        details:
          additionalProperties: {}
          description: |-
            Additional machine-readable context, present for some codes:
            `lastServer` has `actorTypes` (array of strings);
            `noHostsReachable` has `errors` (array of `HostError`)
          type: object
        message:
          description: A human-readable description of the error
          type: string
        requestId:
          description: The request ID, also returned in the `X-Request-Id` header
          type: string
        retryable:
          description: Present and `true` when repeating the request later may succeed
          type: boolean
      required:
      - code
      - message
      type: object
...
    Job:
      properties:
        actorId:
          type: string
        actorType:
          type: string
        attempts:
          description: The number of attempts, present only for terminal jobs (`completed`,
            `dead`)
          type: integer
        createdAt:
          description: Omitted when unknown
          format: date-time
          type: string
        cron:
          description: The cron expression, for cron jobs
          type: string
        dueTime:
          format: date-time
          type: string
        endedAt:
          description: When the job ended, only for terminal jobs and omitted when
            unknown
          format: date-time
          type: string
        interval:
          description: The repeat interval, for repeating jobs
          type: string
        jobId:
          type: string
        lastError:
          description: The last error, only for terminal jobs and omitted when empty
          type: string
        method:
          type: string
        status:
          description: '`dead` is dead-lettered'
          enum:
          - pending
          - active
          - completed
          - dead
          type: string
        workflow:
          allOf:
          - $ref: '#/components/schemas/WorkflowLink'
          description: Present for jobs of workflow actors
      required:
      - actorId
      - actorType
      - dueTime
      - jobId
      - method
      - status
      type: object
...
    Page-Alarm:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Alarm'
          type: array
        nextCursor:
          description: Opaque cursor to pass back as `cursor` to fetch the next page,
            omitted on the last page
          type: string
      required:
      - items
      type: object
...
    Page-Job:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Job'
          type: array
        nextCursor:
          description: Opaque cursor to pass back as `cursor` to fetch the next page,
            omitted on the last page
          type: string
      required:
      - items
      type: object
...
    WorkflowLink:
      description: Links an actor of a workflow to its workflow instance, derived
        from the actor type and ID naming conventions.
      properties:
        capability:
          description: The capability of a worker or undo actor type, when it has
            one
          type: string
        instanceId:
          description: The workflow instance, for orchestrator, worker, and undo actors
            whose ID could be parsed
          type: string
        role:
          enum:
          - orchestrator
          - worker
          - undo
          - registry
          - other
          type: string
        step:
          description: The step, for worker and undo actors
          type: string
        taskIndex:
          description: The task index within the step, for worker and undo actors
          type: integer
        workflow:
          type: string
      required:
      - role
      - workflow
      type: object
...
  securitySchemes:
    bearerAuth:
      description: 'A read-only or management token from the server configuration,
        sent as `Authorization: Bearer <token>`. Read-only tokens grant every scope
        except those ending in `:manage`; management tokens grant every scope.'
      in: header
      name: Authorization
      type: apiKey
info:
  contact: {}
  description: |
    The management API exposes read-only views of a Francis cluster (hosts, runtimes, placements, in-memory activations, actor state, jobs, alarms, and workflows), plus a small set of audited actions (drain a host, deactivate an actor, cancel/suspend/resume a workflow instance).

    ## Authentication and scopes

    Every route under `/api/v1` requires an `Authorization: Bearer <token>` header.
    `GET /api/v1/openapi.yaml` and `GET /healthz` (served at the root, outside `/api/v1`) are public and don't require a token.

    Each operation declares the scope it requires in its description and in the `x-required-scope` extension.
    The server is configured with two token lists:

    - **Read-only tokens** receive every scope except those ending in `:manage`.
    - **Management tokens** receive every scope.

    | Scope | Grants |
    | --- | --- |
    | `cluster:read` | Cluster summary, runtimes, hosts |
    | `actors:read` | Activations, placements, actor types, actor-state listings |
    | `actors:state:read` | Reading the stored state of an actor (audited); a workflow's actor types also need `workflows:data:read` |
    | `workflows:read` | Workflows, instances, event history |
    | `workflows:data:read` | Including workflow instance input and output in instance details, and reading the stored state of a workflow's actor types (audited) |
    | `jobs:read` | Jobs and alarms |
    | `actors:manage` | Deactivating an actor |
...

...

...

...

...

...

...

...

...

...

...
openapi: 3.0.3
paths:
...
  /api/v1/alarms:
    get:
      description: |-
        Requires scope `jobs:read`.

        Lists alarms, ordered by actor type, actor ID, and alarm name.
        The filters can be repeated, and the alarms are then listed one combination of their values after another, by actor type and then actor ID, both in alphabetical order.
      operationId: listAlarms
      parameters:
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          type: integer
      - description: Opaque cursor returned as nextCursor by the previous page; omit
          for the first page
        in: query
        name: cursor
        schema:
          type: string
      - description: Only return alarms of these actor types; repeat the parameter
          for several
        in: query
        name: type
        schema:
          items:
            type: string
          type: array
      - description: Only return alarms of these actor IDs, which requires type; repeat
          the parameter for several
        in: query
        name: id
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page-Alarm'
          description: A page of alarms
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: List alarms
      tags:
      - Jobs
      x-required-scope: jobs:read
...
  /api/v1/jobs:
    get:
      description: |-
        Requires scope `jobs:read`.

        Lists durable jobs, ordered by job ID.
        The filters can be repeated, and the jobs are then listed one combination of their values after another: by actor type and actor ID, both in alphabetical order, then by status in the order `pending`, `active`, `completed`, `dead`, and ordered by job ID within each.
        Terminal jobs (`completed`, `dead`) are only listed while a record of them is retained, as configured by the actor type's job retention.
      operationId: listJobs
      parameters:
      - description: Maximum number of items to return
        in: query
        name: limit
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          type: integer
      - description: Opaque cursor returned as nextCursor by the previous page; omit
          for the first page
        in: query
        name: cursor
        schema:
          type: string
      - description: Only return jobs of these actor types; repeat the parameter for
          several
        in: query
        name: type
        schema:
          items:
            type: string
          type: array
      - description: Only return jobs of these actor IDs, which requires type; repeat
          the parameter for several
        in: query
        name: id
        schema:
          items:
            type: string
          type: array
      - description: Only return jobs in these statuses; repeat the parameter for
          several
        in: query
        name: status
        schema:
          items:
            enum:
            - pending
            - active
            - completed
            - dead
            type: string
          type: array
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page-Job'
          description: A page of jobs
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: List jobs
      tags:
      - Jobs
      x-required-scope: jobs:read
  /api/v1/jobs/{jobId}:
    get:
      description: Requires scope `jobs:read`.
      operationId: getJob
      parameters:
      - description: The job ID
        in: path
        name: jobId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          description: The job
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`badRequest`: an invalid path segment, query parameter, cursor,
            or request body'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`unauthorized`: the bearer token is missing or unknown'
          headers:
            WWW-Authenticate:
              description: Always Bearer realm="francis-management" when the token
                is missing or unknown
              schema:
                type: string
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`forbidden`: the token does not grant the scope the route
            requires'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`notFound`: the job does not exist'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`internal`: an unexpected server error'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: '`timeout`: the request timed out; retryable'
          headers:
            X-Request-Id:
              description: A unique ID assigned to the request, also returned as requestId
                in error bodies and recorded in audit logs
              schema:
                type: string
      security:
      - bearerAuth: []
      summary: Get a job
      tags:
      - Jobs
      x-required-scope: jobs:read
...